Add FieldSeeker layer
The plan is to keep ArcGIS simple, a thin layer over API requests. I'll
then build richer structures on top of it for FieldSeeker, a product
built on ArcGIS.

With this I can at least get the source value for a Service Request,
barely. There are many improvements to be made, but I'm calling this a
good checkpoint.