Switch error detection so that errors.Is(err, ...) works
Without this we end up with a problem of instantiating a different error
to check against because we're using

errors.Is(err, &InvalidatedRefreshTokenError{})

which doesn't work the way we want unless we pack it with all the
information the client shouldn't have to know like the code and the
error message. Instead we'll use constants to reference against so now
we do

errors.Is(err, argcis.InvalidatedRefreshTokenError)

which works correctly.