Switch from implicit layer IDs
Instead we'll keep a dynamic mapping internally of layer IDs and have an
explicit mapping from layer names to an enum. This allows us to have a
known set of layers that we recognize which we can map to structs.

This also removes the ability to do raw queries to clients - we're going
to mediate these queries through functions instead.